refactor: modifico la ubicación del endpoint de consulta de reservas para admins
refactor: Se remueve endpoint de consulta de reservas de BookingController a ReportController
