+++ src/main/java/com/rkmd/toki_no_nagare/controller/report/ReportController.java
@@ -3,16 +3,14 @@ package com.rkmd.toki_no_nagare.controller.report;
 import com.rkmd.toki_no_nagare.dto.booking.BookingResponseDto;
 import com.rkmd.toki_no_nagare.service.BookingService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 
 @RestController
 @RequestMapping("")
-public class ReportController {
+public class ReportController implements ReportControllerResources{
 
     @Autowired
     private BookingService bookingService;
@@ -22,5 +20,12 @@ public class ReportController {
         BookingResponseDto booking = bookingService.getBookingByCode(codeId);
 
         return ResponseEntity.ok().body(booking);
+    }
+
+
+    @GetMapping(value = "/v1/reports/bookings/{bookingCode}/contacts/{dni}", produces = "application/json")
+    @ResponseStatus(value = HttpStatus.OK)
+    public BookingResponseDto getBookingByCodeAndDni(String bookingCode, Long dni) {
+        return bookingService.getBookingByCodeAndDni(bookingCode, dni);
     }
 }
