+++ src/main/java/com/rkmd/toki_no_nagare/controller/report/ReportControllerResources.java
@@ -1,42 +1,18 @@
-package com.rkmd.toki_no_nagare.controller.booking;
+package com.rkmd.toki_no_nagare.controller.report;
 
-import com.rkmd.toki_no_nagare.dto.booking.CreateBookingRequestDto;
-import com.rkmd.toki_no_nagare.dto.booking.CreateBookingResponseDto;
-import com.rkmd.toki_no_nagare.entities.booking.BookingStatus;
+import com.rkmd.toki_no_nagare.dto.booking.BookingResponseDto;
 import com.rkmd.toki_no_nagare.exception.ApiError;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 
-@Tag(name = "Reservas")
-public interface BookingControllerResources {
+public interface ReportControllerResources {
 
   @Operation(
-      summary = "Crea una reserva",
-      description = "Endpoint para la web de RKMD",
-      responses = {
-          @ApiResponse(
-              responseCode = "201",
-              description = "Created",
-              useReturnTypeSchema = true),
-          @ApiResponse(
-              responseCode = "400",
-              description = "Bad Request",
-              content = @Content(schema = @Schema(implementation = ApiError.class))),
-          @ApiResponse(
-              responseCode = "500",
-              description = "Internal server error",
-              content = @Content(schema = @Schema(implementation = ApiError.class)))})
-  CreateBookingResponseDto createBooking(@RequestBody CreateBookingRequestDto request);
-
-
-  @Operation(
-      summary = "Obtiene una reserva según el código de reserva.",
+      summary = "Obtiene una reserva según el código y el dni.",
       description = "Endpoint para la web de RKMD",
       responses = {
           @ApiResponse(
@@ -51,6 +27,26 @@ public interface BookingControllerResources {
               responseCode = "500",
               description = "Internal server error",
               content = @Content(schema = @Schema(implementation = ApiError.class)))})
-  ResponseEntity<BookingStatus> getBookingStatus(@PathVariable("code_id") String codeId);
+  BookingResponseDto getBookingByCodeAndDni(@PathVariable("bookingCode") String bookingCode, @PathVariable("dni") Long dni);
+
+
+  @Operation(
+      summary = "Obtiene los datos de una reserva según el código.",
+      description = "Endpoint para la web de RKMD",
+      responses = {
+          @ApiResponse(
+              responseCode = "200",
+              description = "Ok",
+              useReturnTypeSchema = true),
+          @ApiResponse(
+              responseCode = "404",
+              description = "Not Found",
+              content = @Content(schema = @Schema(implementation = ApiError.class))),
+          @ApiResponse(
+              responseCode = "500",
+              description = "Internal server error",
+              content = @Content(schema = @Schema(implementation = ApiError.class)))})
+  ResponseEntity<BookingResponseDto> getBookingReport(@PathVariable("code_id") String codeId);
+
 
 }
